+++ yirgacheffe/_backends/enumeration.py
@@ -55,6 +55,7 @@ class operators(Enum):
     LSHIFT = 47
     RSHIFT = 48
     ASAREA = 49
+    ASPROJECTION = 50
 
 class dtype(Enum):
     """Represents the type of data returned by a layer.