+++ yirgacheffe/_backends/mlx.py
@@ -213,6 +213,9 @@ def astype_op(data, datatype):
 def as_area_op(data, new_area): # pylint: disable=W0613
     return data
 
+def as_projection_op(data, projection, method): # pylint: disable=W0613
+    return data
+
 operator_map: dict[op, Callable] = {
     op.ADD: mx.array.__add__,
     op.SUB: mx.array.__sub__,
@@ -262,4 +265,5 @@ operator_map: dict[op, Callable] = {
     op.LSHIFT: mx.array.__lshift__,
     op.RSHIFT: mx.array.__rshift__,
     op.ASAREA: as_area_op,
+    op.ASPROJECTION: as_projection_op,
 }